feat(cache): split runtime memory feature
