+++ Cargo.toml
@@ -108,7 +108,7 @@ rustfs-notify = { path = "crates/notify", version = "1.0.0-beta.12" }
 rustfs-io-metrics = { path = "crates/io-metrics", version = "1.0.0-beta.12" }
 rustfs-io-core = { path = "crates/io-core", version = "1.0.0-beta.12" }
 rustfs-object-capacity = { path = "crates/object-capacity", version = "1.0.0-beta.12" }
-rustfs-object-data-cache = { path = "crates/object-data-cache", version = "1.0.0-beta.12" }
+rustfs-object-data-cache = { path = "crates/object-data-cache", version = "1.0.0-beta.12", default-features = false }
 rustfs-log-analyzer = { path = "crates/log-analyzer", version = "1.0.0-beta.12" }
 rustfs-obs = { path = "crates/obs", version = "1.0.0-beta.12" }
 rustfs-policy = { path = "crates/policy", version = "1.0.0-beta.12" }
